+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/ImageMode/annotations/ImageAnnotations.ts
@@ -10,20 +10,13 @@ import { PinholeCameraModel } from "@foxglove/den/image";
 import { ImageAnnotations as FoxgloveImageAnnotations } from "@foxglove/schemas";
 import { MessageEvent, SettingsTreeAction, Topic } from "@foxglove/studio";
 import { normalizeAnnotations } from "@foxglove/studio-base/panels/Image/lib/normalizeAnnotations";
-import {
-  Annotation as NormalizedAnnotation,
-  PointsAnnotation as NormalizedPointsAnnotation,
-} from "@foxglove/studio-base/panels/Image/types";
 import { ImageModeConfig } from "@foxglove/studio-base/panels/ThreeDeeRender/IRenderer";
-import { SRGBToLinear } from "@foxglove/studio-base/panels/ThreeDeeRender/color";
 
-import { DynamicBufferGeometry } from "../../../DynamicBufferGeometry";
+import { TopicAnnotationsRenderable } from "./TopicAnnotationsRenderable";
 import { SettingsTreeEntry } from "../../../SettingsManager";
 import { IMAGE_ANNOTATIONS_DATATYPES } from "../../../foxglove";
 import { IMAGE_MARKER_ARRAY_DATATYPES, IMAGE_MARKER_DATATYPES } from "../../../ros";
 import { topicIsConvertibleToSchema } from "../../../topicIsConvertibleToSchema";
-
-const tempVec3 = new THREE.Vector3();
 
 interface ImageAnnotationsContext {
   topics(): readonly Topic[];
@@ -34,116 +27,6 @@ interface ImageAnnotationsContext {
     schemaNames: Set<string>,
     handler: (messageEvent: MessageEvent<T>) => void,
   ): void;
-}
-
-class PointsRenderable extends THREE.Object3D {
-  #geometry: DynamicBufferGeometry;
-  #points: THREE.Points;
-  #pointsMaterial: THREE.PointsMaterial;
-
-  public constructor() {
-    super();
-    this.#geometry = new DynamicBufferGeometry();
-    this.#geometry.createAttribute("position", Float32Array, 3);
-    this.#geometry.createAttribute("color", Uint8Array, 4, true);
-    this.#pointsMaterial = new THREE.PointsMaterial({
-      size: 1,
-      sizeAttenuation: false, // Point size is based on viewport size and image size, not 3D position
-      vertexColors: true,
-    });
-    this.#points = new THREE.Points(this.#geometry, this.#pointsMaterial);
-    this.add(this.#points);
-  }
-
-  public dispose() {
-    this.#geometry.dispose();
-    this.#pointsMaterial.dispose();
-  }
-
-  public update(
-    annotation: NormalizedPointsAnnotation & { style: "points" },
-    cameraModel: PinholeCameraModel,
-    scale: number,
-  ): void {
-    const { points, outlineColors, outlineColor, fillColor, thickness } = annotation;
-    this.#geometry.resize(points.length);
-    const positionAttribute = this.#geometry.getAttribute("position") as THREE.BufferAttribute;
-    const colorAttribute = this.#geometry.getAttribute("color") as THREE.BufferAttribute;
-    const positions = positionAttribute.array as Float32Array;
-    const colors = colorAttribute.array as Uint8Array;
-    const fallbackColor = outlineColor && outlineColor.a > 0 ? outlineColor : fillColor;
-
-    this.#pointsMaterial.size = thickness * scale;
-    for (let i = 0; i < points.length; i++) {
-      const color = outlineColors[i] ?? fallbackColor;
-      const point = points[i]!;
-      if (cameraModel.projectPixelTo3dPlane(tempVec3, point)) {
-        positions[i * 3 + 0] = tempVec3.x;
-        positions[i * 3 + 1] = tempVec3.y;
-        positions[i * 3 + 2] = tempVec3.z;
-        colors[i * 4 + 0] = SRGBToLinear(color?.r ?? 0) * 255;
-        colors[i * 4 + 1] = SRGBToLinear(color?.g ?? 0) * 255;
-        colors[i * 4 + 2] = SRGBToLinear(color?.b ?? 0) * 255;
-        colors[i * 4 + 3] = (color?.a ?? 0) * 255;
-      } else {
-        positions[i * 3 + 0] = NaN;
-        positions[i * 3 + 1] = NaN;
-        positions[i * 3 + 2] = NaN;
-      }
-    }
-    positionAttribute.needsUpdate = true;
-    colorAttribute.needsUpdate = true;
-  }
-}
-
-class TopicAnnotationsRenderable extends THREE.Object3D {
-  #points?: PointsRenderable;
-  #scale = 1;
-
-  public dispose(): void {
-    this.#points?.dispose();
-  }
-
-  public updateScale(scale: number): void {
-    this.#scale = scale;
-    //TODO: re-render immediately with new scale
-  }
-
-  public update(annotations: NormalizedAnnotation[], cameraModel: PinholeCameraModel) {
-    if (this.#points) {
-      this.#points.visible = false;
-    }
-    for (const annotation of annotations) {
-      switch (annotation.type) {
-        case "circle":
-          break;
-        case "points":
-          switch (annotation.style) {
-            case "points":
-              if (!this.#points) {
-                this.#points = new PointsRenderable();
-                this.add(this.#points);
-              }
-              this.#points.visible = true;
-              this.#points.update(
-                annotation as typeof annotation & { style: typeof annotation.style },
-                cameraModel,
-                this.#scale,
-              );
-              break;
-            case "polygon":
-              break;
-            case "line_strip":
-              break;
-            case "line_list":
-              break;
-          }
-          break;
-        case "text":
-          break;
-      }
-    }
-  }
 }
 
 /**
